version with poliastro
In this version I use propagators from poliastro. There are a lot of prints, time in debris file and memory usage trackers in the pipeline. I was using it to track what's going on and what is the most consuming. In the pipeline I was calculating magnitudes. It turned out, the poliastro propagator is very slow.